+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added the `cloney.start` command to start a new Cloney project.
 - Added missing property `variable.validate` to the `.cloney.yaml` metadata file schema.
+- The extension now searches for the nearest `.cloney.yaml` file relative to the current file's path instead of always searching from the workspace root. This allows for multiple Cloney projects to be used in the same workspace.
 
 ## 0.0.2